+++ src/main/ts/core/OptionResolver.types.ts
