+++ src/main/ts/core/Tools.ts
